refactor: hide server url and personal information using local properties
- subtract interceptor class from ProductClient
