+++ app/src/main/java/woowacourse/shopping/data/database/client/ProductClient.kt
@@ -1,17 +1,15 @@
 package woowacourse.shopping.data.database.client
 
-import android.util.Base64
 import com.google.gson.GsonBuilder
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+import woowacourse.shopping.BuildConfig
 import woowacourse.shopping.data.model.dto.coupon.CouponDto
 import woowacourse.shopping.domain.service.RetrofitService
 
 object ProductClient {
-    private const val BASE_URL = "http://54.180.95.212:8080"
+    private const val BASE_URL = BuildConfig.BASE_URL
 
     private val gsonCouponParser = GsonBuilder()
         .registerTypeAdapter(CouponDto::class.java, CouponDeserializer())
@@ -24,42 +22,11 @@ object ProductClient {
             .client(provideOkHttpClient(AppInterceptor()))
             .build()
 
-    val service = client.create(RetrofitService::class.java)
+    val service: RetrofitService = client.create(RetrofitService::class.java)
 
     private fun provideOkHttpClient(interceptor: AppInterceptor): OkHttpClient =
         OkHttpClient.Builder().run {
             addInterceptor(interceptor)
             build()
         }
-
-    class AppInterceptor : Interceptor {
-        private val email = "junyoung-won"
-        private val password = "password"
-
-        override fun intercept(chain: Interceptor.Chain): Response {
-            val headString = createAuthorizationHeaderString(email, password)
-            val request =
-                chain.request()
-                    .newBuilder()
-                    .addHeader(AUTHORIZATION_HEADER, headString)
-                    .build()
-            return chain.proceed(request)
-        }
-
-        private fun createAuthorizationHeaderString(
-            email: String,
-            password: String,
-        ): String {
-            val authString = "$email:$password"
-
-            val encodedAuthString =
-                Base64.encodeToString(authString.toByteArray(Charsets.UTF_8), Base64.NO_WRAP)
-
-            return "Basic $encodedAuthString"
-        }
-
-        companion object {
-            const val AUTHORIZATION_HEADER = "Authorization"
-        }
-    }
 }
